pyproject.toml
[project]
name = "crawl-topcv-jobs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "bs4>=0.0.2",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "ipykernel>=6.29.5",
    "lxml>=6.0.1",
    "matplotlib>=3.7.5",
    "nbconvert>=7.16.6",
    "openpyxl>=3.1.5",
    "pandas>=2.0.3",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "selenium>=4.27.1",
    "webdriver-manager>=4.0.2",
]
